+++ src/angular-cli/shared/helpers/angular-cli.ts
@@ -1,0 +1,7 @@
+import * as vscode from 'vscode';
+
+export class AngularCliHelper {
+    public static runNgCliCommand(terminal: vscode.Terminal, command: string) {
+        terminal.sendText(`npm run ng -- ${command}`);
+    }
+}