+++ src/global/interfaces/commands-container.ts
@@ -1,0 +1,5 @@
+import * as vscode from 'vscode';
+
+export interface CommandsContainer {
+    registerCommands(context: vscode.ExtensionContext, namespace: string, options: { [key: string]: any });
+}