+++ src/global/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './module-with-commands';
+export * from './commands-container';