+++ src/global/models/command-args.ts
@@ -1,0 +1,7 @@
+export class CommandArgs {
+    fsPath: string;
+    external: string;
+    $mid: number;
+    path: string;
+    scheme: string;
+}