+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Change Log
-All notable changes to the "hw" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Added
+- Create a spec file for Angular component
+- Extract unit test for method
+
+## [0.0.3] - 2017-03-20
+### Added
+- Basic unit test support (mark a method, right-click -> Add test method)
+
+## [0.0.2] - 2017-03-19
+### Added
+- Support creating Angular components (right-click on a folder -> New component)
+
+## [0.0.1]
 - Initial release